[FEAT] Improve the 'monitoring' demo (#184)
Add a short text in the page to explain what the example shows.

fix: don't reload the frequency diagram on every toggle
fix: display links to source code

Reduce execution time displayed in edge overlays - only use seconds. This
simulates the latency of the bpmn engine (time taken to process subsequent
element of the process instance) which should be close to 0.