[FEAT] 'Predictions' demo: reduce visibility of already executed paths
Style the paths for the 2 use cases. We use very specific css selectors (at
least more precise than in already existing examples and demo).
It is also applied to the label but is only visible after zooming or panning
because of a bug in bpmn-visualization.

Improve the layout of the pizza diagram:
  - remove msg flow icon (not possible to style for now with the current
  version of bpmn-visualization)
  - align elements in pool to have vertical msg flow (improve clarity)