[REFACTOR] Remove duplications in comparison with other BPMN vendors
Introduce shared classes for library and UI management + a 'state' shared object.
Examples extend them which reduces duplications a lot.

The shared classes allow a clearer separation between the libs management and
actions on the UI.
Errors are better handled: for instance, the 'Last loaded file' info is no
longer updated on load failure.

kie-editors-standalone example: manage the 'Fit View' button title and text
based on the 'fitView' state for a better feedback for the user.