[FEAT] Add a vanilla TypeScript project built with Vite
The project is created from the Vite vanilla-ts template. It loads a BPMN
diagram and applies CSS classes to a Task.

Also update the TS project built with rollup
  - lint
  - no need to generate declaration file (not used as a lib)
  - activate "strict" and clean options (already managed by the strict flag, or
  unnecessary)
  - bump TS to 4.5, required as we use new 4.5 type import

Add a GitHub workflow that checks that the TypeScript projects build. It
currently contains a hack for the bpmn-visualization package that doesn't
correctly bundle the TS definitions.