[FEAT] Disable pool collapsing (#307)
The upcoming bpmn-visualization version changes the edge rendering. There is no more usage of the Segment Connector.
This setting was recomputing the whole edge path. It was able to set the terminal waypoints to the collapsed pools when
the edge was targeting a inner element of the pool (hidden when the pool is collapsed).
There is no more waypoints' computation with the new implementation: in this case, we see a dangling edge. The edge
terminal waypoint are inside the pool.