[INFRA] Improve the bpmn-visualization update workflow (#322)
Apply changes recently tested in the playground repository.

The major change is the usage of `GH_RELEASE_TOKEN` instead of `GITHUB_TOKEN` with the dawidd6/action-download-artifact.
According to the action documentation, this lets the action access to artifacts stored in another repository. As we know
this is working with more restricted `GITHUB_TOKEN` permissions in the playground repository, let's apply the change now
instead of waiting for an issue when the permissions restriction will be in place in the `examples` repository.