[INFRA] Update devDependencies in projects
Apply to all projects except Parcel TS as dependencies were already up-to-date.

In addition, configure npm to not generate package-lock. We don't commit lock files because we don't want to have to
update it when bumping the bpmn-visualization version. So it is not used for CI tasks. It was previously only generated
locally and generally not used by developers.