[FEAT] Simplify integration of bpmn-visualization in TS projects
Remove the additional TypeScript configuration that was needed prior version 0.27.0. The README has been updated to warn
TS user that use older version.

The parcel project now directly calls the mxGraph code in addition to the vite project. This provides more test for this
use case and better validated the npm package.