[INFRA] Publish PR preview on surge.sh for TypeScript projects (#438)
Generate a home page that lists all built projects with both the last
released and development version of `bpmn-visualization`.

Update the configuration of the projects to ensure they can be deployed
in non root context (i.e. not in `/`) and that are all built in the same
directory (to apply the same processing to generate the preview
resources).