fix(monitoring-demo-frequency): update the stroke of the overlays
The stroke color should be the same as the fill color to demonstrate a non default stroke color (as already done in the
time use case).
It had been put in black by mistake during a refactoring released with version 0.19.2.
The stroke now always has the right color.

Also update the JSDoc (parameter types) of functions involved in the fix.