feat: check integrity when loading `bpmn-visualization` from CDN (#541)
Previously, we load the `bpmn-visualization` IIFE bundle from the
`jsdelivr` CDN without integrity. This was not consistent with the rest
of resources that we load from this CDN and this could lead to safety
risks.
The version of `bpmn-visualization` is updated automatically when a new
release is available. Changing the version is easy, but the integrity
computation requires more work. As a first implementation, we decided to
not do it.

Now, the integrity string is first computed from the bundle of the npm
package (the package is downloaded by the GitHub workflow). Then, all
pages using the bundle from the jsdelivr CDN are updated with the new
version (as before) and with the new integrity string.
All pages have been updated to use the `bpmn-visualization` 0.39.0
integrity value in order to benefit from the improvement without waiting
for the new version.